+++ common/src/main/java/net/mehvahdjukaar/supplementaries/client/particles/RotationTrailEmitter.java
@@ -33,7 +33,8 @@ public class RotationTrailEmitter extends NoRenderParticle {
             this.level.addParticle(ModParticles.ROTATION_TRAIL.get(),
                     this.x, this.y, this.z,
                     this.axis.get3DDataValue() * ccw,
-                    this.radius, (i * 90 + 45) + ccw * timeSinceStart * RotationTrailParticle.SPEED);
+                    this.radius,
+                    (i * 90 + 45) + ccw * timeSinceStart * RotationTrailParticle.SPEED);
         }
 
         ++this.timeSinceStart;