+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/entities/trades/StructureMapListing.java
@@ -24,7 +24,8 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Optional;
 import java.util.function.Function;
 
-public record StructureMapListing(Item cost, int minPrice, int maxPrice, Optional<ItemCost> cost2, HolderSet<Structure> structure,
+public record StructureMapListing(Item cost, int minPrice, int maxPrice, Optional<ItemCost> cost2,
+                                  HolderSet<Structure> structure,
                                   int maxTrades, float priceMult, int level,
                                   String mapName, int mapColor,
                                   ResourceLocation mapMarker) implements ModItemListing {
@@ -41,20 +42,17 @@ public record StructureMapListing(Item cost, int minPrice, int maxPrice, Optiona
                     ExtraCodecs.POSITIVE_INT.optionalFieldOf("price_max", 13).forGetter(StructureMapListing::maxPrice),
                     ItemCost.CODEC.optionalFieldOf("price_secondary").forGetter(StructureMapListing::cost2),
                     TARGET_CODEC.fieldOf("structure").forGetter(p -> p.structure),
-                    ExtraCodecs.POSITIVE_INT.optionalFieldOf( "max_trades", 16).forGetter(StructureMapListing::maxTrades),
-                    ExtraCodecs.POSITIVE_FLOAT.optionalFieldOf( "price_multiplier", 0.05f).forGetter(StructureMapListing::priceMult),
-                    Codec.intRange(1, 5).optionalFieldOf( "level", 1).forGetter(StructureMapListing::level),
-                    Codec.STRING.optionalFieldOf( "map_name", "").forGetter(p -> p.mapName),
-                    ColorUtils.CODEC.optionalFieldOf( "map_color", 0xffffff).forGetter(p -> p.mapColor),
-                    ResourceLocation.CODEC.optionalFieldOf( "map_marker", ResourceLocation.parse("")).forGetter(p -> p.mapMarker))
+                    ExtraCodecs.POSITIVE_INT.optionalFieldOf("max_trades", 16).forGetter(StructureMapListing::maxTrades),
+                    ExtraCodecs.POSITIVE_FLOAT.optionalFieldOf("price_multiplier", 0.05f).forGetter(StructureMapListing::priceMult),
+                    Codec.intRange(1, 5).optionalFieldOf("level", 1).forGetter(StructureMapListing::level),
+                    Codec.STRING.optionalFieldOf("map_name", "").forGetter(p -> p.mapName),
+                    ColorUtils.CODEC.optionalFieldOf("map_color", 0xffffff).forGetter(p -> p.mapColor),
+                    ResourceLocation.CODEC.optionalFieldOf("map_marker", ResourceLocation.parse("")).forGetter(p -> p.mapMarker))
             .apply(i, StructureMapListing::new));
-
-
 
 
     @Override
     public MerchantOffer getOffer(@NotNull Entity entity, @NotNull RandomSource random) {
-
         ItemStack itemstack = AdventurerMapsHandler.createCustomMapForTrade(entity.level(), entity.blockPosition(),
                 structure, mapName.isEmpty() ? null : mapName, mapColor, mapMarker.getPath().isEmpty() ? null : mapMarker);
         if (itemstack.isEmpty()) return null;