+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/utils/ItemsUtil.java
@@ -1,18 +1,23 @@
 package net.mehvahdjukaar.supplementaries.common.utils;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.mehvahdjukaar.moonlight.api.item.additional_placements.AdditionalItemPlacement;
 import net.mehvahdjukaar.moonlight.api.item.additional_placements.AdditionalItemPlacementsAPI;
 import net.mehvahdjukaar.moonlight.api.item.additional_placements.BlockPlacerItem;
+import net.mehvahdjukaar.supplementaries.SuppPlatformStuff;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.KeyLockableTile;
+import net.mehvahdjukaar.supplementaries.common.events.overrides.InteractEventsHandler;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.food.FoodProperties;
 import net.minecraft.world.inventory.ClickAction;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.BlockItem;
@@ -25,6 +30,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
@@ -48,13 +54,32 @@ public class ItemsUtil {
 
     //helper for slingshot. Calls both block item and this in case it as additional behavior
     public static InteractionResult place(Item item, BlockPlaceContext pContext) {
-        //this also calls mixin
-        //TODO: find better way to do
-        if (item instanceof BlockItem bi) return bi.place(pContext);
+
+        Player player = pContext.getPlayer();
+        ItemStack stack = pContext.getItemInHand();
+        Level level = pContext.getLevel();
+        BlockHitResult hit = pContext.getHitResult();
+
+        //block override. mimic forge event that would have called these
+        //we cant call all forge events unfortunately otherwise we could alter the world as if player was there and not just place extra blocks
+        InteractionResult overrideResult = InteractEventsHandler.onItemUsedOnBlockHP(player, level, stack, InteractionHand.MAIN_HAND, hit);
+        if (overrideResult.consumesAction()) {
+            return overrideResult;
+        }
+        overrideResult = InteractEventsHandler.onItemUsedOnBlock(player, level, stack, InteractionHand.MAIN_HAND, hit);
+        if (overrideResult.consumesAction()) {
+            return overrideResult;
+        }
+
+        //this also would be called from useOn. Since we aren't calling use on we must call it directly
         var placement = AdditionalItemPlacementsAPI.getBehavior(item);
         if (placement != null) return placement.overridePlace(pContext);
+        if (item instanceof BlockItem bi) {
+            return SuppPlatformStuff.placeBlockItem(bi,pContext);
+        }
         return InteractionResult.PASS;
     }
+
 
     public static void addStackToExisting(Player player, ItemStack stack, boolean avoidHands) {
         var inv = player.getInventory();
@@ -116,7 +141,7 @@ public class ItemsUtil {
 
     //TODO: move to lib
     @ExpectPlatform
-    public static ItemStack tryExtractingItem(Level level, @Nullable Direction dir, BlockPos pos, @Nullable  BlockEntity container) {
+    public static ItemStack tryExtractingItem(Level level, @Nullable Direction dir, BlockPos pos, @Nullable BlockEntity container) {
         throw new AssertionError();
     }